Present synchronized secrets in SDS
This commit changes Network Policy parsing so that, when secret
synchronization for Network Policy is enabled, referenced
Kubernetes Secrets will be sent using references to SDS secrets,
along with ensuring that synchronized secrets are read into
SDS correctly.

When secret sync for Policy is disabled, the UseFullContext fix
option will work to fix a bug as per the fix for #31761. When
secret sync is enabled, this option is ignored, as SDS secrets
are not susceptible to the same issue.

This change also keeps the existing option to read Secret values
referenced in Network Policy TLS Secrets from the filesystem
rather than Kubernetes. In this case, the secret values will
be sent inline in NPDS.

Currently includes a nasty hack to get around race conditions
between operator secret sync and agent needing the synced secret.

Signed-off-by: Nick Young <nick@isovalent.com>