Fix incorrect Gateway API processing for same-name GAMMA services
This commit fixes a bug with Gateway API processing that can occur
when:

* A GAMMA HTTPRoute references a Service by name - for example, a
  Service named `foo` exists in the `foo` namespace and is referenced
  by a HTTPRoute.
* A Gateway exists in the same namespace with the same name - in this
  example, a `foo` Gateway _also_ exists in the `foo` namespace.

In this case, the GAMMA HTTPRoute's config would be incorrectly added
to the Gateway's generated CiliumEnvoyConfig.

This is caused by the lack of a check for the Gateway Kind and Group
in the Gateway reconciler when checking to see if HTTPRoutes are
relevant for that Gateway.

This commit adds those checks, and a test to ensure we don't do this
again.

Signed-off-by: Nick Young <nick@isovalent.com>